Fixed firestore database fetching
I made such a silly mistake in fetching my data from my database the path had an extra space in it which I had not noticed. Firebase is very sensitive when it comes to that kind of thing which is why it had not worked thus far. Also why I had not committed anything in a while.

Updated the Front screens so now they will present with the data from their respective categories.